Migrate backend from SQLite to PostgreSQL
Switch the backend to use PostgreSQL: update Backend/.env.example to use DATABASE_URL and make prisma.config.ts read DATABASE_URL. Replace the SQLite Prisma datasource with the postgresql provider in prisma/schema.prisma and adjust generated timestamp defaults to use extract(epoch from now()). Update package.json to use @prisma/adapter-pg and add pg/@types/pg dev deps. Dockerfile was simplified (remove native build toolchain and SQLite volume) and entrypoint docs updated for Postgres. Add a retry loop to docker-entrypoint.sh so `npx prisma db push` waits for the database to be ready. Remove the old data/sqlite ignores from .gitignore. These changes prepare the app and container for running against a PostgreSQL database.

+++ Backend/src/api/tidal.ts
@@ -472,7 +472,7 @@ async function loadLocalPlaylist(
 	const aggregateRows = await prisma.$queryRaw<
 		Array<{ total: number; duration: number }>
 	>`
-		SELECT COUNT(*) as total, COALESCE(SUM(t.duration), 0) as duration
+		SELECT COUNT(*)::int as total, COALESCE(SUM(t.duration), 0)::int as duration
 		 FROM playlist_tracks pt
 		 LEFT JOIN tracks t ON t.id = pt.track_id
 		 WHERE pt.playlist_id = ${playlistId}`;
